+++ pydantic_settings/__init__.py
@@ -1,9 +1,23 @@
 import warnings
 
 from .main import BaseSettings
+from .sources import (
+    DotEnvSettingsSource,
+    EnvSettingsSource,
+    InitSettingsSource,
+    PydanticBaseSettingsSource,
+    SecretsSettingsSource,
+)
 from .version import VERSION
 
-__all__ = ('BaseSettings',)
+__all__ = (
+    'BaseSettings',
+    'PydanticBaseSettingsSource',
+    'InitSettingsSource',
+    'SecretsSettingsSource',
+    'EnvSettingsSource',
+    'DotEnvSettingsSource',
+)
 
 __version__ = VERSION
 warnings.warn(
